Positive Typology Container by type of study and location

Complementary to the Stegomyia indices, the type of positive container is determined to establish the preference of mosquitoes in the use of breeding sites.

Usage

get_typology_container_by_type_of_study_and_location(
  df,
  st = "Verificacion",
  var = c("CIUDAD_OBREGÓN", "HERMOSILLO")
)

Arguments

df

The dataframe with entomology information processed by the function "clean_raw_data".

st

a string of type of study selected, "Encuesta" or "Verificacion". By default, it is set to "Verificacion".

var

receive string of the "Location", which is the geographic variable used to calculate Positive Typology Container.

Value

The dataframe with location of the selection of type of study and geographic variable of location.

Examples

df_tc <-get_typology_container_by_type_of_study_and_location(df_typ,
                                                             st = "Verificacion",
                                                             "CIUDAD_OBREGÓN"
)
#> Warning: Casa_Revisada with 0
head(df_tc)
#>        Localidad Pct_Recipientes_Tratables Pct_Recipientes_Controlables
#> 1 CIUDAD_OBREGÓN                        20                           80
#>   Pct_Recipientes_Eliminables
#> 1                           0